Fix controllers, comment endpoints

+++ app/src/application/interfaces/interactors/feature_interactor.py
@@ -1,5 +1,7 @@
+from sqlalchemy.exc import SQLAlchemyError
+
 from src.domain.dto.feature_dto import BaseFeatureResponse, CreateFeatureRequest, CreateFeatureResponse, DeleteFeatureResponse, GetFeatureResponse, GetAllFeaturesResponse
-from src.application.exceptions import IdNotValidError
+from src.application.exceptions import DatabaseException, IdNotValidError
 from src.application.interfaces.transaction_manager import ITransactionManager
 from src.application.interfaces.repositories import feature_repository
 
@@ -13,13 +15,17 @@ class GetFeatureInteractor:
     async def __call__(self, feature_id: int) -> GetFeatureResponse:
         if feature_id < 1:
             raise IdNotValidError
-        feature = await self._feature_repository.get_feature_by_id(feature_id)
-        
-        return GetFeatureResponse(
-            id=feature.id,
-            name=feature.name,
-            icon_url=feature.icon_url
-        )
+        try:
+            feature = await self._feature_repository.get_feature_by_id(feature_id)
+
+            return GetFeatureResponse(
+                id=feature.id,
+                name=feature.name,
+                icon_url=feature.icon_url
+            )
+
+        except SQLAlchemyError:
+            raise DatabaseException("Failed to get feature by id in db")
 
 
 class GetAllFeaturesInteractor:
@@ -29,16 +35,20 @@ class GetAllFeaturesInteractor:
         self._feature_repository = feature_repository
 
     async def __call__(self) -> GetAllFeaturesResponse:
-        features = await self._feature_repository.get_features()
-        data = [
-            BaseFeatureResponse(
-                id=feature.id,
-                name=feature.name,
-                icon_url=feature.icon_url
-            ) 
-            for feature in features
-        ]
-        return GetAllFeaturesResponse(data=data)
+        try:
+            features = await self._feature_repository.get_features()
+            data = [
+                BaseFeatureResponse(
+                    id=feature.id,
+                    name=feature.name,
+                    icon_url=feature.icon_url
+                ) 
+                for feature in features
+            ]
+            return GetAllFeaturesResponse(data=data)
+
+        except SQLAlchemyError:
+            raise DatabaseException("Failed to get all features in db")
 
 
 class AddFeatureInteractor:
@@ -50,14 +60,18 @@ class AddFeatureInteractor:
         self._transaction_manager = transaction_manager
 
     async def __call__(self, feature_request: CreateFeatureRequest) -> CreateFeatureResponse:
-        feature = await self._feature_repository.add_feature(feature_request)
-        await self._transaction_manager.commit()
+        try:
+            feature = await self._feature_repository.add_feature(feature_request)
+            await self._transaction_manager.commit()
 
-        return CreateFeatureResponse(
-            id=feature.id,
-            name=feature.name,
-            icon_url=feature.icon_url
-        )
+            return CreateFeatureResponse(
+                id=feature.id,
+                name=feature.name,
+                icon_url=feature.icon_url
+            )
+
+        except SQLAlchemyError:
+            raise DatabaseException("Failed to add feature in db")
 
 
 class DeleteFeatureInteractor:
@@ -69,8 +83,15 @@ class DeleteFeatureInteractor:
         self._transaction_manager = transaction_manager
 
     async def __call__(self, feature_id: int) -> DeleteFeatureResponse:
-        feature = await self._feature_repository.get_feature_by_id(feature_id)
-        feature_response = await self._feature_repository.delete_feature(feature)
-        await self._transaction_manager.commit()
+        if feature_id < 1:
+            raise IdNotValidError
 
-        return feature_response
+        try:
+            feature = await self._feature_repository.get_feature_by_id(feature_id)
+            feature_response = await self._feature_repository.delete_feature(feature)
+            await self._transaction_manager.commit()
+
+            return feature_response
+
+        except SQLAlchemyError:
+            raise DatabaseException("Failed to add feature in db")
